# aether-commerce/docs/openapi/aether.v1.yaml
openapi: 3.0.3
info:
  title: Aether API
  version: 1.0.0
  description: Worker API for the Aether premium commerce demo.
servers:
  - url: https://api.aether.example/api/v1
  - url: http://localhost:8787/api/v1
paths:
...
  /products/{id}:
    get:
      tags: [Public]
      summary: Product by public ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Product }
  /products/slug/{slug}:
    get:
      tags: [Public]
      summary: Product by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200": { description: Product }
...
  /categories/{slug}/products:
    get:
      tags: [Public]
      summary: Products by category
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200": { description: Product collection }
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Private admin dashboard
      security:
        - bearerAuth: []
      responses:
        "200": { description: Dashboard }
  /admin/products:
    get:
      tags: [Admin]
      summary: Admin product list
      security:
        - bearerAuth: []
      responses:
        "200": { description: Products }
  /admin/orders:
    get:
      tags: [Admin]
      summary: Admin order list
      security:
        - bearerAuth: []
      responses:
        "200": { description: Orders }
  /admin/summary:
    get:
      summary: Private admin summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Private summary
        "403":
          $ref: "#/components/responses/Error"
  /admin/products/{id}/override:
    patch:
      summary: Update product override
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Override saved
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
...